% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_columns_cat}
\alias{adm.check_columns_cat}
\title{Function to check if all columns listed in categories are also in the dataset}
\usage{
adm.check_columns_cat(datafile, categories = NULL)
}
\arguments{
\item{datafile}{data input}

\item{categories}{categories dataframe}
}
\description{
Function to check if all columns listed in categories are also in the dataset
}
