% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_columns_var}
\alias{adm.check_columns_var}
\title{Function to check if all columns listed in variables are also in the dataset and vice-versa}
\usage{
adm.check_columns_var(datafile, variables)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check if all columns listed in variables are also in the dataset and vice-versa
}
