% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_date}
\alias{adm.check_date}
\title{Function to check date}
\usage{
adm.check_date(datafile, variables, ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check date
}
