% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_datetime}
\alias{adm.check_datetime}
\title{Function to check datetime}
\usage{
adm.check_datetime(datafile, variables, ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check datetime
}
