% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_encrypted_columns}
\alias{adm.check_encrypted_columns}
\title{Function to check encryption of columns}
\usage{
adm.check_encrypted_columns(datafile, variables)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check encryption of columns
}
