% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_encrypted_values}
\alias{adm.check_encrypted_values}
\title{Function to check encryption of values}
\usage{
adm.check_encrypted_values(datafile, variables)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check encryption of values
}
