% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_entitytype}
\alias{adm.check_entitytype}
\title{Function to check entity type}
\usage{
adm.check_entitytype(variables)
}
\arguments{
\item{variables}{variables dataframe}
}
\description{
Function to check entity type
}
