% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_ids}
\alias{adm.check_ids}
\title{Function to check if datafile has duplicated ids}
\usage{
adm.check_ids(datafile, id.name = "id", ...)
}
\arguments{
\item{datafile}{data input}

\item{id}{The name of the column representing the entity identifiers. Default is 'id'.}
}
\description{
Function to check if datafile has duplicated ids
}
