% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_infinite}
\alias{adm.check_infinite}
\title{Function to check on Inf values}
\usage{
adm.check_infinite(datafile, ...)
}
\arguments{
\item{datafile}{data input}
}
\description{
Function to check on Inf values
}
