% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_labels}
\alias{adm.check_labels}
\title{Function to check presence of labels & descriptions in variables}
\usage{
adm.check_labels(variables)
}
\arguments{
\item{variables}{variables dataframe}
}
\description{
Function to check presence of labels & descriptions in variables
}
