% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.check_minmax}
\alias{adm.check_minmax}
\title{Function to check data on minimum & maximum values if defined in variables}
\usage{
adm.check_minmax(datafile, variables, categories = NULL)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}
}
\description{
Function to check data on minimum & maximum values if defined in variables
}
