% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{adm.complete_diffdf}
\alias{adm.complete_diffdf}
\title{Function to check datafile, variables & categories in one go}
\usage{
adm.complete_diffdf(datalist1, datalist2, ...)
}
\arguments{
\item{datalist1}{List object: list(datafile = datafile, categories = categories, variables = variables)}

\item{datalist2}{List object: list(datafile = datafile, categories = categories, variables = variables)}
}
\value{
findings, dataframe with differences (if there are any)
}
\description{
Function to check datafile, variables & categories in one go
}
