% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{adm.fix_valuetype}
\alias{adm.fix_valuetype}
\title{Fix data variable value types to match Opal expectations}
\usage{
adm.fix_valuetype(datafile, variables)
}
\arguments{
\item{datafile}{A data frame or list containing the data to be checked}

\item{variables}{A data frame or list with 'name' and 'valueType' columns}
}
\value{
A modified version of the variables list with corrected value types
}
\description{
Compares the expected value types from a variables list with the actual types
in the data file. If mismatches are found, updates the value type in the variables
list and prints a message.
}
