% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{adm.remove_haven}
\alias{adm.remove_haven}
\title{Remove haven labels and missing definitions}
\usage{
adm.remove_haven(x)
}
\arguments{
\item{x}{A vector, of class `haven_labelled` or `haven_labelled_spss`.}
}
\value{
The input vector with haven-specific attributes and classes removed.
}
\description{
Strips all haven-specific attributes from a labelled vector.
}
