% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{adm.run_all_checks}
\alias{adm.run_all_checks}
\title{Function to check run all checks}
\usage{
adm.run_all_checks(datafile, variables, categories = NULL)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}
}
\description{
Function to check run all checks
}
