% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{adm.set_method}
\alias{adm.set_method}
\title{Function to set method for force & overwrite}
\usage{
adm.set_method(method)
}
\arguments{
\item{method}{String: write, update or overwrite}
}
\value{
method, vector with settings for force & overwrite
}
\description{
Function to set method for force & overwrite
}
