% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{adm.table_copy}
\alias{adm.table_copy}
\title{Function to copy a table within Opal}
\usage{
adm.table_copy(opal, projname, tablenames, ...)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{projname}{Origin opal project name}

\item{tablename}{Origin opal table name}
}
\description{
Function to copy a table within Opal
}
