% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{adm.table_save_diffdf}
\alias{adm.table_save_diffdf}
\title{Function to save a table from R to Opal with a diffdf check}
\usage{
adm.table_save_diffdf(
  opal,
  project,
  table,
  datafile,
  variables = NULL,
  categories = NULL,
  method = "write",
  path = NULL,
  max_retries = 3,
  ...
)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{project}{Origin opal project name}

\item{table}{Origin opal table name}

\item{datafile}{data dataframe}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}

\item{method}{String: write, update or overwrite}

\item{path}{String: Path of folder}

\item{max_retries}{Integer: The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\description{
Function to save a table from R to Opal with a diffdf check
}
