% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tres.R
\name{adm.tres_connection}
\alias{adm.tres_connection}
\title{Function to check tres connection}
\usage{
adm.tres_connection(connection, ...)
}
\arguments{
\item{connection}{Connection with RTRES}
}
\description{
Function to check tres connection
}
