% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tres.R
\name{adm.tres_decryption}
\alias{adm.tres_decryption}
\title{Function to decrypt with tres}
\usage{
adm.tres_decryption(connection, datafile, columns = NULL, ...)
}
\arguments{
\item{connection}{Connection with RTRES}

\item{datafile}{data input}

\item{columns}{selected columns for decryption}
}
\value{
datafile, decrypted datafile
}
\description{
Function to decrypt with tres
}
