% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tres.R
\name{adm.tres_encryption}
\alias{adm.tres_encryption}
\title{Encrypt Data Using TRES}
\usage{
adm.tres_encryption(connection, values, search_image = FALSE, ...)
}
\arguments{
\item{connection}{A connection object used to communicate with the TRES service.}

\item{values}{Column values to encrypt.}

\item{search_image}{A logical value indicating whether to search for image data after encryption.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
Result as encrypted values
}
\description{
Encrypts data using the TRES encryption service.
}
