% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{adm.view_create}
\alias{adm.view_create}
\title{Create or Update a View in OPAL with Dictionary and Permissions}
\usage{
adm.view_create(
  opal,
  project,
  table,
  source,
  variables,
  categories = NULL,
  ...
)
}
\arguments{
\item{opal}{A connection object to the OPAL server.}

\item{project}{The name of the project where the view is to be created or updated.}

\item{table}{The name of the table (view) to be created or updated.}

\item{source}{The source table(s) used to define the view.}

\item{variables}{A data frame containing variable definitions (name, type, etc.).}

\item{categories}{Optional data frame of category mappings for variables.}
}
\value{
A logical value indicating whether the view was created or updated.
}
\description{
This function creates or updates a view in an OPAL project based on a source table.
It checks if the view exists, creates or updates it, sets the script column,
updates the dictionary entries, and removes user permissions.
}
