% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{adm.write_to_excel}
\alias{adm.write_to_excel}
\title{Function to write diffdf to excel}
\usage{
adm.write_to_excel(df_list, path, ...)
}
\arguments{
\item{df_list}{list of dataframes from diffdf}

\item{path}{path to output dir}
}
\description{
Function to write diffdf to excel
}
