% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{capture_logs}
\alias{capture_logs}
\title{Function to capture all warnings & errors when running all check functions}
\usage{
capture_logs(run_function)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to capture all warnings & errors when running all check functions
}
