% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.cat_labels}
\alias{check.cat_labels}
\title{Check Categorical Labels Completeness}
\usage{
check.cat_labels(datafile, categories, ...)
}
\arguments{
\item{datafile}{A data frame containing the actual data.}

\item{categories}{A data frame or list containing category definitions.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A warning if any value in a categorical column is missing from the category list.
}
\description{
Ensures that all values in categorical columns are present in the category definitions.
}
