% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.cat_text_labels}
\alias{check.cat_text_labels}
\title{Function to check if datafile values are present in categorie labels (text values only)}
\usage{
check.cat_text_labels(datafile, categories, ...)
}
\arguments{
\item{datafile}{data input}

\item{categories}{categories dataframe}
}
\description{
Function to check if datafile values are present in categorie labels (text values only)
}
