% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.character_ids}
\alias{check.character_ids}
\title{Check if ID Column is Character Type}
\usage{
check.character_ids(datafile, id.name = "id", ...)
}
\arguments{
\item{datafile}{A data frame containing the actual data.}

\item{id.name}{A character string specifying the ID column name (default: "id").}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A warning if the ID column is not of character type.
}
\description{
Ensures that the ID column is of character type.
}
