% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.columns_cat}
\alias{check.columns_cat}
\title{Check Column Mismatch Between Datafile and Categories}
\usage{
check.columns_cat(datafile, categories = NULL, ...)
}
\arguments{
\item{datafile}{A data frame containing the actual data.}

\item{categories}{A data frame or list containing category definitions.}
}
\value{
A list of warnings about missing columns in categories.
}
\description{
Validates that all columns in the categories object are present in the datafile.
}
