% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.date}
\alias{check.date}
\title{Function to check date}
\usage{
check.date(datafile, variables, format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}

\item{format}{format of dates (default = "\%Y-\%m-\%d")}
}
\description{
Function to check date
}
