% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.datetime}
\alias{check.datetime}
\title{Function to check datetime}
\usage{
check.datetime(datafile, variables, ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check datetime
}
