% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.duplicated_ids}
\alias{check.duplicated_ids}
\title{Check for Duplicated IDs in Data}
\usage{
check.duplicated_ids(datafile, id.name = "id", ...)
}
\arguments{
\item{datafile}{A data frame containing the actual data.}

\item{id.name}{A character string specifying the ID column name (default: "id").}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A warning if duplicated IDs are found.
}
\description{
Ensures that the ID column contains no duplicated values.
}
