% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.duplicated_rows}
\alias{check.duplicated_rows}
\title{Function to check for duplicated rows in variables & categories}
\usage{
check.duplicated_rows(variables, categories = NULL, ...)
}
\arguments{
\item{variables}{variables dataframe}

\item{categories}{categories dataframe}
}
\description{
Function to check for duplicated rows in variables & categories
}
