% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.encrypted_values}
\alias{check.encrypted_values}
\title{Function to check encryption of values}
\usage{
check.encrypted_values(datafile, variables, ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check encryption of values
}
