% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.entitytype}
\alias{check.entitytype}
\title{Function to check entity type}
\usage{
check.entitytype(variables, ...)
}
\arguments{
\item{variables}{variables dataframe}
}
\description{
Function to check entity type
}
