% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.ids}
\alias{check.ids}
\title{Function to check if datafile has duplicated ids}
\usage{
check.ids(datafile, id.name = "id", ...)
}
\arguments{
\item{datafile}{data input}

\item{id.name}{The name of the column representing the entity identifiers. Default is 'id'.}
}
\description{
Function to check if datafile has duplicated ids
}
