% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.infinite}
\alias{check.infinite}
\title{Function to check on Inf values}
\usage{
check.infinite(datafile, ...)
}
\arguments{
\item{datafile}{data input}
}
\description{
Function to check on Inf values
}
