% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.run_all}
\alias{check.run_all}
\title{Function to check run all checks}
\usage{
check.run_all(datafile, variables, categories = NULL)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}
}
\description{
Function to check run all checks
}
