% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.valuetype}
\alias{check.valuetype}
\title{Check Value Type Consistency Between Data and Variables}
\usage{
check.valuetype(datafile, variables, ...)
}
\arguments{
\item{datafile}{A data frame containing the actual data.}

\item{variables}{A data frame or list containing variable definitions.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A list of warnings about mismatched value types.
}
\description{
Compares the data type of each column in the datafile with the specified value types in variables.
}
