% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.var_column_datatype}
\alias{check.var_column_datatype}
\title{Function to check if all columns in variables are character type}
\usage{
check.var_column_datatype(variables, ...)
}
\arguments{
\item{variables}{variables dataframe}
}
\description{
Function to check if all columns in variables are character type
}
