% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_check_categoriesminmax_generic.R
\name{check_categoriesminmax_generic}
\alias{check_categoriesminmax_generic}
\title{Check if values in the datafile are conform the min and max in the var dictionary}
\usage{
check_categoriesminmax_generic(
  datafiles,
  vars,
  cats = NULL,
  key = "id",
  report_path = NULL,
  report_name = "Report",
  silent = FALSE,
  conditional = NULL,
  includeNA = FALSE
)
}
\arguments{
\item{datafiles}{list. A list containing for each element a datafile}

\item{vars}{list. A list containing for each element a var dictionary, that matches the datafile from datafiles with the same element}

\item{cats}{an optional list. A list containing for each element a cat dictionary, that matches the datafile from datafiles with the same element. If NULL, than only the min/max can be checked and not the categories.}

\item{key}{string. The key of the datafile that will be read into opal.}

\item{report_path}{string. Folder where the report needs to be saved. If \code{NULL} the report is not saved but only returned.}

\item{silent}{boolean. Whether or not to silence the printing of messages while running the function (FALSE is default).}

\item{conditional}{an optional list. In each list element for a specific variable an extra filter statement can be included. Should be constructed as: `variable_name ~ statement`, where the variable name must match a variable in the dataset and the statement should be an evaluable statement by the dplyr::filter function.}

\item{includeNA}{boolean. Whether or not to include the NAs from the min_max report}
}
\value{
a list with an element for each supplied datafile. Per list element there can be two separate list elements containing matrices with the findings. Only present if there are non-zero findings:
* min_max A matrix with 5 columns and rows equal to the number of findings. Columns are `{key}`, `variable name`, `value`, `min` and `max`. Value is the observed value that is outside the min/max range
* outside_cat A matrix with 4 column and rows equal to the number of findings. Columns are `{key}`, `variable name`, `value` and `categories`. Value is the observed value, categories the current known labels
}
\description{
In the var dictionary it is possible to supply a min and a max, to give the boundaries of the values. However, it is still possible that values are observed that are outside this boundary and it can be important to check if this is the case.
}
\note{
Cannot be used reliably when the variables have labels, so when a table is retrieved from opal
}
\author{
Lars van der Burg & Thekla Jansen
}
