% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_check_diffdf_opal_generic.R
\name{check_diffdf_opal_generic}
\alias{check_diffdf_opal_generic}
\title{Create diffdf overviews and a report (optionally in excel/csv). This function uses all the code from the script \code{Check_diffdf_opal_generic.R} with some small adjustments.}
\usage{
check_diffdf_opal_generic(
  datafile = NULL,
  datafile2 = NULL,
  var = NULL,
  var2 = NULL,
  cat = NULL,
  cat2 = NULL,
  comparison = "mod",
  comp_key = "id",
  date_time_format = c("Ymd HMS", "Ymd"),
  report_path = NULL,
  report_name = "Report",
  report_title = "",
  suppress_warnings = FALSE,
  opt_repl_castor = FALSE,
  opt_rm_VarDiff_null = FALSE,
  aggregate_VarDiff = FALSE,
  opt_calc_VarDiff_spaces = TRUE,
  opt_calc_VarDiff_diff = TRUE,
  opt_rm_VarDiff_diff_0 = FALSE,
  report_out = "single",
  manual_csv = FALSE,
  opt_rm_spaces_around = FALSE,
  opt_rm_AttribD_null = FALSE,
  opt_rm_nonUTF8 = FALSE
)
}
\arguments{
\item{datafile}{tibble. The datafile, prior to uploading to opal. In general the BASE for the diffdf comparison.}

\item{datafile2}{tibble. The datafile retrieved from opal. In general the COMPARE for the diffdf comparison.}

\item{var}{tibble. Variables of datafile. In general the BASE for the diffdf comparison.}

\item{var2}{tibble. Variables of datafile2, retrieved from opal. In general the COMPARE for the diffdf comparison.}

\item{cat}{tibble. Categories of datafile. In general the BASE for the diffdf comparison.}

\item{cat2}{tibble. Categories of datafile2, retrieved from opal. In general the COMPARE for the diffdf comparison.}

\item{comparison}{string. Which comparison to run, there are three options: c("base", "mod", "both"). The base comparison compares the unadjusted datafile and dictionary, the mod comparison (default) compares the modified datafile and dictionary and the both comparison performes both.}

\item{comp_key}{string. The key at which the datafile and datafile2 are compared, must be present in both}

\item{date_time_format}{vector. To calculate a difference between two (date) time variables. Need to know the possible formats these (date) times are written down.}

\item{report_path}{string. Folder where the report needs to be saved. If \code{NULL} the report is not saved but only returned.}

\item{report_name}{string. Name of the report to save. A sys.Date() is always added to the report_name.}

\item{report_title}{string. The title of the comparison}

\item{suppress_warnings}{logical. Whether to suppress warnings in the \code{diffdf} function, is a \code{diffdf} argument (default = FALSE)}

\item{opt_repl_castor}{logical. Whether to replace a `;` with a `-`, and replace `\\n` and `\\r` with a "". This is what is done in the excel export of Castor. Note that ALL ; will be replaced by a -}

\item{opt_rm_VarDiff_null}{logical. Whether to remove all VarDiff differences with only ""/NA/"null" (default = FALSE)}

\item{aggregate_VarDiff}{logical. Whether or not to make a aggregated VarDiff}

\item{opt_calc_VarDiff_spaces}{logical. Whether to try to calculate if the difference between BASE and COMPARE is due to whitespaces at the start and/or end of a string. If with stringr::str_trim(..) there is no difference between BASE and COMPARE, difference is set to 0 for that row}

\item{opt_calc_VarDiff_diff}{logical. Whether to try to calculate differences between BASE and COMPARE for all VarDiff rows (default = TRUE)}

\item{opt_rm_VarDiff_diff_0}{logical. Whether to remove all rows with VarDiff difference of 0 (default = FALSE)
## The following four arguments are eligible for discarding, dont think they will be used anymore (report_out = "single" is then always the case)}

\item{report_out}{string. In what format the report needs to be saved, there are two options: c("single", "multiple"). Single exports all files in one excel, multiple saves a separate file for each component.}

\item{manual_csv}{logical. Whether to save the VarDiff as .csv document. Required when because of too many strange symbols the \code{read_excel} function doesnot work (default = FALSE)}

\item{opt_rm_AttribD_null}{logical. Whether to remove attribute differences with only ""/null/NULL/NA differences (default = FALSE)}

\item{opt_rm_nonUTF8}{logical. Whether first to remove all non-UTF8 characters and then to remove all start and trailing spaces and all double spaces (default = FALSE)}
}
\value{
A diffdf report is made (and optionally saved) that is a list with at most 7 elements: NumDiff, VarDiff, AttribD, VarClas, VarMode, ExtRows, ExtCols. Each element contains information about differences between the datafiles, vars and cats. If a file is empty, it is omitted from the saving of the plot.
* VarMode: Difference modes for each variable or attribute, e.g., character vs numeric. If different, variables are not checked! But only an issue when table = diff_xmod
* NumDiff: All differences that there are on a variable level, e.g., 2700 value mismatches
* VarDiff: All mismatches that are denoted in NumDiff, e.g., for all 2700 observations the value in the BASE and the value in the COMPARE
* AttribD: All attribute differences per variable, e.g., different entity_type or Min/Max values. Here opal.repeatable is important! Null=0 is fine
* VarClas: Difference classes for each variable or attribute, e.g., haven_labelled character vs character
* ExtRows: Are there extra rows added?
* ExtCols: Are there extra columns added? e.g., label vs label:en
}
\description{
Create a report with the differences between two opal datasets. Usually one that will uploaded to opal and the same dataset that is retrieved from opal
}
\details{
In total 6 diffdf comparisons are made: datafile-datafile2; var-var2; cat-cat2; datafilemod-datafile2mod; varmod-var2mod and cat-cat2mod. Compared to the supplied tibbles, some small adjustments are made to the ...mod versions. The last three diffdfs are added to check some regular problems (e.g., changed variable class). Because of these adjustments, differences will of course be found. However, because one can see the type of difference and the origen of the difference, one can say whether it is a problem. Additionally, because of these adjustments, one can check whether the differences is created (not in the supplied tibble, but only in the ...mod version) or solved (present in the supplied tibble, but not in the ...mod version). This thus gives some insights into the severity of the difference.
}
\author{
Lars van der Burg
}
