% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_checks_opal_R.R
\name{checks_opal_R}
\alias{checks_opal_R}
\title{Function that performs some general checks}
\usage{
checks_opal_R(
  datafile,
  var,
  cat = NULL,
  key = "id",
  min_max = FALSE,
  silent = FALSE
)
}
\arguments{
\item{datafile}{data.frame. Containing the datafile with for each column the variables described by \code{var}. The datafile that will be supplied to \code{opalr::dictionary.apply()}.}

\item{var}{data.frame. Containing the variable dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{cat}{data.frame. Containing the category dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{key}{string. The key of the datafile that will be read into opal.}

\item{min_max}{boolean. Is there a min/max in the datafile that needs to be checked (FALSE is default).}

\item{silent}{boolean. Whether or not to silence the printing of a short conclusion for each check (FALSE is default).}
}
\value{
A list with a data.frame with the found issues and optionally (if min_max = TRUE) with the issues found with the check_categoriesmin_max function. Additionally, for every check a text is printed indicating whether or not an issue is found.
}
\description{
Function that performs various checks on the datafile and dictionary, essential if data has to be put into opal. Not all detected issues are an actual problem, that is up to the data manager to sort out. But it is a handy start.
}
\details{
The following checks are executed:
1. if there are duplicated columns.
2. If there are duplicated keys.
3. If there are variables in the var dictionary that are not in the datafile.
4. If there are variables in the cat dictionary that are not in the datafile.
5. If there are variables that are not described in the var dictionary.
6. If there is an entityType in the var dictionary it must be unique, absent is allowed.
7. If all valueTypes are opal-compatible.
8. If all repeatable values are valid
9. If there is a label or description in the var dictionary.
10. If there is an encrypted column in the var dictionary
10.5 If the column exist, check if "yes" and "SI" are actually encrypted
11. If all essential cat columns are present
12.5 If all missing values are valid
13. If there are Inf values in the datafile, need to be changed to NA to save in opal
14. If all observed values adhere to the min/max in the var dictionary (optional).
15. If all date and datetimes have a format that is accepted by opal.

The last check (15. if the date and datetimes are opal-compatible) is sub-optimal. It is very difficult to get a simple but complete check for all date(times) formats in opal. So interpret this with caution. Issues might be not a problem and no detected issues doesn't mean that everything is perfect.

Previous checks include:
- Number of decimals. If there are values with more than decimals, currently not possible with opal. In opal 5 this problem is no more
}
\author{
Lars van der Burg
}
