% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_datafile_conform_var_change.R
\name{datafile_conform_var_change}
\alias{datafile_conform_var_change}
\title{Function to check whether all datafile variable classes are conform the valueType in the dictionary, if not, datafile class is changed}
\usage{
datafile_conform_var_change(
  datafile,
  var,
  date_time_format = c("Ymd", "Ymd HMS")
)
}
\arguments{
\item{datafile}{data.frame. Containing the dataset with for each column the variables described by var. The datafile that will be supplied to opalr::dictionary.apply().}

\item{var}{data.frame.Containing per row the description of each variable. The dictionary that will be supplied to opalr::dictionary.apply().}

\item{date_time_format}{character. Specifying the date and/or datetime format used in the datafile. Multiple options are possible, it will then look for the option that best fits the data. Here we can ignore the \-, \/, \% sign and other signs in the notation. Default = c("Ymd", "Ymd HMS")}
}
\value{
A list is returned with two elements. Element 1 is the altered datafile of the same dimensions as the supplied datafile. Element 2 is a matrix containing the remaining problems, i.e., the variables for which a class mismatch is found but no solution is currently supplied or for which there is data loss due to the class conversion. This matrix has a row for each issue and 3 columns:
* name: containing the corresponding variables,
* issue: the issue with the variable and
* info: with information about the issue.
}
\description{
We check in the var dictionary whether the specified valueType matches the actual class of the variable in the datafile. If this is not the case, OPAL will guess this automatically possibly without giving you a warning. Some combinations are correct (e.g., R: integer -> Opal: integer), others can be incorrect (e.g., R: character -> Opal: integer) due to various reasons. For certain combinations the class of the datafile is changed, for others only a message is given indicating that there might be a problem that needs to be checked out. When the class is changed, it is possible that some data is lost. A message is given when this is the case, denoting for which variable and for which conversion data loss takes place.
}
\author{
Lars van der Burg
}
