% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_datafile_conform_var_check.R
\name{datafile_conform_var_check}
\alias{datafile_conform_var_check}
\title{Function to check whether all datafile variable classes are conform the valueType in the dictionary.}
\usage{
datafile_conform_var_check(datafile, var, silent = FALSE)
}
\arguments{
\item{datafile}{data.frame. Containing the dataset with for each column the variables described by \code{var}. The datafile that will be supplied to \code{opalr::dictionary.apply()}.}

\item{var}{data.frame.Containing per row the description of each variable. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{silent}{boolean. Whether or not to silence the printing of a short conclusion for each check (FALSE is default).}
}
\value{
A matrix with rows equal to the amount of variables for which there is a data issue and three rows:
* name: indicating the variable with the data issue
* issue: the data issue
* info: information about the data issue
}
\description{
In this function we check which cases are not conform the valueTypes acceptable for opal, and whether the classes of the datafile are conform the specified valueTypes, if you want to adjust them, used \code{datafile_conform_var_check}
}
\examples{
var = var |> mutate(valueType = ifelse(valueType == "integer", str_to_title(valueType), valueType))
datafile_conform_var_check(datafile, var)

}
\author{
Lars van der Burg
}
