% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_decrypt_data.R
\name{decrypt_data}
\alias{decrypt_data}
\title{Decrypt a vector or a tibble/data.frame}
\usage{
decrypt_data(con, data, vars_to_decrypt = NULL)
}
\arguments{
\item{con}{string. Tres connection as obtained via \code{tres_connect}}

\item{data}{vector or tibble/data.frame. The data that needs to be decrypted}

\item{vars_to_decrypt}{optional vector. Indicates which variables need to be decrypted if \code{data} is a tibble or a data.frame. If NULL all variables will be decrypted}
}
\value{
decrypted vector or a tibble with the vars_to_decrypt as decrypted. Also print a summary of the warnings that the \code{tres_decrypt} function returned.
}
\description{
wrapper around the zorgTTP rtres package to decrypt data. Added a check for connection, summary of the warnings and allows for decryption of a vector or a tibble/data.frame
}
\note{
When decrypt_data is used inside a dplyr::mutate call the summary of the warnings will be cumulative. Unclear how to reset dplyr::last_dplyr_warnings
}
\author{
Lars van der Burg
}
