% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_delete_table_opal.R
\name{delete_table_opal}
\alias{delete_table_opal}
\title{Function to delete a table from opal}
\usage{
delete_table_opal(opal, projname, tablename, ...)
}
\arguments{
\item{opal}{string. A working opalr::opal_login}

\item{projname}{string. Origin opal project name}

\item{tablename}{string. Origin opal table name}
}
\description{
To delete a table from opal. Because of sensitivity of operation, an extra confirmation is required.
}
\examples{
It is possible to also delete multiple tables at once:
tables <- c("TABLE1", "TABLE2", "TABLE3")
for(table in tables){
  delete_table_opal(opal, projname, table)
}

}
\author{
Lars van der Burg & Thekla Jansen
}
