% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.clean_NA}
\alias{.clean_NA}
\title{Clean data for missing values}
\usage{
.clean_NA(datafile)
}
\arguments{
\item{datafile}{A vector or data frame containing data to clean}
}
\value{
A cleaned version of the input data with missing values standardized
}
\description{
Replaces empty strings, "NA", and "null" with NA values in a diffdf.
}
