% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.set_method}
\alias{.set_method}
\title{Function to set method for force & overwrite}
\usage{
.set_method(method)
}
\arguments{
\item{method}{String: write, update or overwrite}
}
\value{
method, vector with settings for force & overwrite
}
\description{
Function to set method for force & overwrite
}
