% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.tres_response_codes}
\alias{.tres_response_codes}
\title{Function to get tres response codes}
\usage{
.tres_response_codes(response)
}
\arguments{
\item{response}{respons codes captures from TRES}
}
\description{
Function to get tres response codes
}
