% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.write_to_excel}
\alias{.write_to_excel}
\title{Function to write diffdf to excel}
\usage{
.write_to_excel(findings, path, ...)
}
\arguments{
\item{findings}{list of dataframes from diffdf}

\item{path}{path to output dir}
}
\description{
Function to write diffdf to excel
}
