% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.write_to_rdata}
\alias{.write_to_rdata}
\title{Write findings to RData file}
\usage{
.write_to_rdata(findings, path, ...)
}
\arguments{
\item{findings}{A list of findings, with keys corresponding to object types (e.g., "datafile", "variables")}

\item{path}{The file path where the RData file will be saved}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
NULL (workbook is saved to path)
}
\description{
Creates RData file and writes each findings object (datafile, variables, categories)
into separate worksheets. If no issues are found, a placeholder message is written.
}
