% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_encrypt_data.R
\name{encrypt_data}
\alias{encrypt_data}
\title{Encrypt a vector or a tibble/data.frame}
\usage{
encrypt_data(
  con,
  data,
  vars_to_encrypt = NULL,
  search_image = FALSE,
  salted_encryption = FALSE
)
}
\arguments{
\item{con}{string. Tres connection as obtained via \code{tres_connect}}

\item{data}{vector or tibble/data.frame. The data that needs to be encrypted}

\item{vars_to_encrypt}{optional vector. Indicates which variables need to be encrypted if \code{data} is a tibble or a data.frame. If NULL all variables will be encrypted}

\item{search_image}{boolean. Whether or not the search_image needs to be included when encrypting (default = TRUE). If TRUE, the search_image needs to be extracted via vec_extract_search_image. Cannot be TRUE together with \code{salted_encryption}.}

\item{salted_encryption}{boolean. Whether or not ... (default = FALSE). Cannot be TRUE together with \code{search_image}.}
}
\value{
encrypted vector or a tibble with the vars_to_encrypt as encrypted. Also print a summary of the warnings that the \code{tres_encrypt} function returned.
}
\description{
wrapper around the zorgTTP rtres package to encrypt data. Added a check for connection, summary of the warnings and allows for encryption of a vector or a tibble/data.frame
}
\note{
When encrypt_data is used inside a dplyr::mutate call the summary of the warnings will be cumulative. Unclear how to reset dplyr::last_dplyr_warnings
}
\author{
Lars van der Burg
}
