% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_import_copy_table_opal.R
\name{import_copy_table_opal}
\alias{import_copy_table_opal}
\title{Copy an existing opal table to a new location in opal}
\usage{
import_copy_table_opal(
  opal,
  projname,
  tablename,
  opal2 = NULL,
  projname2 = NULL,
  tablename2 = NULL,
  ent = NULL,
  report_path = FALSE,
  report_name = "Report",
  comparison = "mod",
  id.name = "id",
  max_tries = 3,
  ...
)
}
\arguments{
\item{opal}{string. A working opalr::opal_login.}

\item{projname}{string. Origin opal project name.}

\item{tablename}{string. Origin opal table name.}

\item{opal2}{optional string. A working opalr::opal_login. If absent, will use `opal`.}

\item{projname2}{string. Destination opal project name. If absent, will use `projname`.}

\item{tablename2}{string. Destination opal table name. If absent, will use `tablename`.}

\item{ent}{string. EntityType used in the datafile. If absent, will use entityType from the imported var dictionary.}

\item{report_path}{string. Indicating where to save the diffdf report. If NULL (default) report will only be returned.}

\item{report_name}{sring. The name of the report for saving. A sys.Date() is always added to the report_name.}

\item{comparison}{string. Which comparison to run, there are three options: c("base", "mod", "both"). The base comparison compares the unadjusted datafile and dictionary, the mod comparison (default) compares the modified datafile and dictionary and the both comparison performes both.}

\item{id.name}{string. The name of the column representing the entity identifiers. Default is 'id'}

\item{max_tries}{integer. The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
An optionally diffdf report. The diffdf report compares the original datafile/dictionary with the newly copied datafile/dictionary in opal.
}
\description{
Because it is not possible to copy a table in opal, the table is imported from opal into R and than written to the new location in opal
}
\author{
Lars van der Burg, Thekla Jansen & Kristel Schaap
}
