% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_import_copy_table_opal_many.R
\name{import_copy_table_opal_many}
\alias{import_copy_table_opal_many}
\title{Copy many opal tables to opal}
\usage{
import_copy_table_opal_many(
  opal,
  projnames,
  tablenames,
  opal2 = NULL,
  projnames2,
  tablenames2,
  ent = NULL,
  report_path = FALSE,
  report_name = "Report",
  comparison = "mod",
  id.name = "id",
  max_tries = 3,
  ...
)
}
\arguments{
\item{opal}{string. A working opalr::opal_login}

\item{projnames}{vector of string. Origin opal project names for each table. Can be either 1 string (which will then be repeated for each table) or have a string per table}

\item{tablenames}{vector of string. Origin opal table names for each table.}

\item{opal2}{optional string. A working opalr::opal_login. If absent, will use 'opal'.}

\item{projnames2}{vector of string. Destination opal project names for each table. Can be either 1 string (which will then be repeated for each table) or have a string per table}

\item{tablenames2}{vector of string. Destination opal table names for each table.}

\item{ent}{string. EntityType used in the datafile. If absent, will use entityType from the imported var dictionary.}

\item{report_path}{string. Indicating where to save the diffdf report. If NULL (default) report will only be returned.}

\item{report_name}{string. The name of the report for saving. A tablename and sys.Date() is always added to the report_name.}

\item{comparison}{string. Which comparison to run, there are three options: c("base", "mod", "both"). The base comparison compares the unadjusted datafile and dictionary, the mod comparison (default) compares the modified datafile and dictionary and the both comparison performes both.}

\item{id.name}{string. The name of the column representing the entity identifiers. Default is 'id'}

\item{max_tries}{integer. The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
An optionally diffdf report for each copied table. The diffdf report compares the original datafile/dictionary with the newly copied datafile/dictionary in opal.
}
\description{
Wrapper for \code{import_copy_table_opal}. If not all tables can be moved (for example, because there is already a table in the new location), that table will be skipped
}
\author{
Lars van der Burg
}
