% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_import_create_table_opal.R
\name{import_create_table_opal}
\alias{import_create_table_opal}
\title{Create a new table or update an existing table in opal}
\usage{
import_create_table_opal(
  opal,
  projname,
  tablename,
  datafile,
  var,
  cat = NULL,
  ent = NULL,
  action = "write",
  id.name = "id",
  report_path = FALSE,
  report_name = "Report",
  comparison = "mod",
  max_tries = 3,
  ...
)
}
\arguments{
\item{opal}{string. A working opalr::opal_login}

\item{projname}{string. Origin opal project name}

\item{tablename}{string. Origin opal table name}

\item{datafile}{tibble. Containing the dataset with for each column the variables described by \code{var}. The datafile that will be supplied to \code{opalr::dictionary.apply()}.}

\item{var}{tibble. Containing the variable dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{cat}{tibble. Containing the category dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{ent}{string. EntityType used in the datafile.}

\item{action}{string. Whether or not just to save the datafile (write, is the default), update an existing table (update) or overwrite an existing table (overwrite).}

\item{id.name}{string. The name of the column representing the entity identifiers. Default is 'id'}

\item{report_path}{string. Indicating where to save the diffdf report. If NULL (default) report will only be returned.}

\item{report_name}{sring. The name of the report for saving. A sys.Date() is always added to the report_name.}

\item{comparison}{string. Which comparison to run, there are three options: c("base", "mod", "both"). The base comparison compares the unadjusted datafile and dictionary, the mod comparison (default) compares the modified datafile and dictionary and the both comparison performes both.}

\item{max_tries}{integer. The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
An optionally diffdf report. The diffdf report is either between the supplied datafile/dictionary and the imported datafile/dictionary from opal or between the old and updated table from opal.
}
\description{
Create a new table in opal and afterwards make the comparison between the supplied datafile and dictionary with the datafile and dictionary imported from opal. If a table is updated, the diffdf comparison will be between the old and new version of the opal table.
}
\author{
Lars van der Burg, Thekla Jansen & Kristel Schaap
}
