% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_import_table_opal2R.R
\name{import_table_opal2R}
\alias{import_table_opal2R}
\title{Function to read a table from opal to R}
\usage{
import_table_opal2R(opal, projname, tablename, id.name = "id", max_tries = 3)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{projname}{Origin opal project name}

\item{tablename}{Origin opal table name}

\item{id.name}{string. The name of the column representing the entity identifiers. Default is 'id'}

\item{max_tries}{integer. The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
a list with the datafile and dictionary as retrieved from opal and the datafile and dictionary updated for use in R
}
\description{
Function that makes an opal ready for use in R. For this also some attributes/characteristics of the datafile and dictionary need to be altered
}
\author{
Lars van der Burg
}
