% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_make_opal_view.R
\name{make_opal_view}
\alias{make_opal_view}
\title{Making and updating views in opal}
\usage{
make_opal_view(
  opal,
  projname,
  tablename,
  opal_view = NULL,
  projname_view = NULL,
  tablename_view,
  var,
  cat = NULL,
  update = FALSE,
  comp_key = "id",
  ent = NULL,
  source = NULL,
  EntityFilter = NULL,
  report_path = FALSE,
  report_name = "Report",
  comparison = "mod",
  max_tries = 3,
  ...
)
}
\arguments{
\item{opal}{string. A working opalr::opal_login where the original table is saved}

\item{projname}{string. Opal project name of the original table}

\item{tablename}{string. Opal table name of the original table}

\item{opal_view}{string. A working opalr::opal_login where the view will be or is saved}

\item{projname_view}{string. Opal project name of the view}

\item{tablename_view}{string. Opal table name of the view}

\item{var}{tibble. Var dictionary to assign to the view}

\item{cat}{tibble. Cat dictionary to assign to the view}

\item{update}{boolean. Whether or not the function should update or make a view with the supplied var and cat dictionary}

\item{comp_key}{string. The name of the column representing the entity identifiers. Default is 'id'}

\item{ent}{string. EntityType used in the datafile.}

\item{source}{list. List of the table names that will be referred to by the view. If NULL, will be combination of projname and tablename}

\item{EntityFilter}{string. The additional filter supplied to the view. Can be build up in two ways: the required format: $this('group').eq('A') or a convenient format: c("group" = "A")}

\item{report_path}{boolean. Indicating where to save the diffdf report. If NULL (default) report will only be returned and not saved.}

\item{report_name}{string. Name of the report to save. A sys.Date() is always added to the report_name.}

\item{max_tries}{integer. The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
Optionally returns a diffdf report comparing the table supplied with `opal`, `projname` and `tablename` with the view supplied with `opal_view`, `projname_view` and `tablename_view`.
}
\description{
Important for making a view in opal is the script column that is made in the var dictionary. Otherwise the view will be empty.
}
\note{
Important that a column script and a column entityType is present in the var dictionary. Otherwise it will not run!
If you want to combine multiple tables in one view you need to add it to the source argument
If you want to have Magma JS in the column script, you need to make the script column before calling this function
Currently (25/06) it is not possible to downsize a view (e.g., from 5 --> 4 variables). Therefore, independent of update, the view will be created from scratch
It is possible that the view breaks down when the original datafile is altered. The view can then not be imported, an error is catched and the requested diffdf is turned off.
For columns in the var dictionary that are not verified the class can only be character, otherwise a parse error will occur
}
\author{
Lars van der Burg
}
