% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_number_datapoints.R
\name{number_datapoints}
\alias{number_datapoints}
\title{Function to count number of datapoints for every column}
\usage{
number_datapoints(datafile, var, cat = NULL, count_Missings = FALSE)
}
\arguments{
\item{datafile}{tibble. Containing the datafile with for each column the variables described by \code{var}. The datafile that will be supplied to \code{opalr::dictionary.apply()}.}

\item{var}{tibble. Containing the variable dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{cat}{tibble. Containing the category dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{count_Missings}{Boolean. Whether or not to count the missings indicated in the cat dictionary as a datapoint (FALSE is default).}
}
\value{
vector called `Mlstr_area::nr_data_points` (name required by MICA) that indicates the number of datapoints per column
}
\description{
For MICA we can indicate the number of actual datapoints for every column in the dataset, can be calculated with this function. Whether or not values that in the cat dictionary are indicated as missing in this calculation can be indicated
}
\author{
Lars van der burg
}
