% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_taxonomy_setup.R
\name{taxonomy_setup}
\alias{taxonomy_setup}
\title{Create YAML file from excel}
\usage{
taxonomy_setup(taxonomy, author = "ADM", license = "CC BY-NC-ND 4.0")
}
\arguments{
\item{taxonomy}{matrix. The taxonomy, obtained from the ADM template}

\item{author}{string. The author of the taxonomy}

\item{license}{string. The license for the taxonomy}
}
\value{
The taxonomy in YAML format
}
\description{
The YAML file can be uploaded into Opal as a taxonomy. The ADM template is required for correct functioning
}
\note{
use writeLines({returned_object}, paste0({path}, ".yml")) to write the YAML file correctly away after running this function
}
\author{
Lars van der Burg
}
