% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_write_table_R2opal.R
\name{write_table_R2opal}
\alias{write_table_R2opal}
\title{Function to write a table from opal to R}
\usage{
write_table_R2opal(
  opal,
  projname,
  tablename,
  datafile,
  var,
  cat = NULL,
  ent = NULL,
  action = "write",
  id.name = "id",
  max_tries = 3,
  ...
)
}
\arguments{
\item{opal}{string. A working opalr::opal_login}

\item{projname}{string. Origin opal project name}

\item{tablename}{string. Origin opal table name}

\item{datafile}{tibble. Containing the dataset with for each column the variables described by \code{var}. The datafile that will be supplied to \code{opalr::dictionary.apply()}.}

\item{var}{tibble. Containing the variable dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{cat}{tibble. Containing the category dictionary for datafile. The dictionary that will be supplied to \code{opalr::dictionary.apply()}.}

\item{ent}{string. EntityType used in the datafile.}

\item{action}{string. Whether or not just to save the datafile (write, is the default), update an existing table (update) or overwrite an existing table (overwrite).}

\item{id.name}{string. The name of the column representing the entity identifiers. Default is 'id'}

\item{max_tries}{integer. The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
list with the datafile and dictionary downloaded from opal after writing it to opal
}
\description{
Function that writes a R table to opal. It is possible to force a table update
}
\note{
Tables can only be written to a project where an user has project permissions. With the creation or updating of a table (independent of chosen options overwrite or force) there will be given table permissions rights for the user to the table. Since the user already has project permissions, these table permissions are discarded.
}
\author{
Lars van der Burg
}
