% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{adm.check_diffdf}
\alias{adm.check_diffdf}
\title{Function to find differences in data}
\usage{
adm.check_diffdf(dataframe1, dataframe2, path = NA, ...)
}
\arguments{
\item{dataframe1}{dataframe}

\item{dataframe2}{dataframe}

\item{path}{path to output dir}
}
\description{
Function to find differences in data
}
