% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{adm.complete_diffdf}
\alias{adm.complete_diffdf}
\title{Compare Multiple Data Frame Lists for Comprehensive Differences}
\usage{
adm.complete_diffdf(datalist1, datalist2, ...)
}
\arguments{
\item{datalist1}{A list containing at least 'datafile' and optionally 'variables' and 'categories'.}

\item{datalist2}{A list containing at least 'datafile' and optionally 'variables' and 'categories'.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A list containing the differences found across datafile, variables, and categories.
}
\description{
This function performs a comprehensive comparison between two lists of data frames, including datafile, variables, and categories.
}
