% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{adm.diffdf}
\alias{adm.diffdf}
\title{Function to find differences in data}
\usage{
adm.diffdf(datafile1, datafile2, keys = NULL, ...)
}
\arguments{
\item{keys}{vector of variables (as strings) that defines a unique row in the base and compare dataframes}

\item{dataframe1}{dataframe}

\item{dataframe2}{dataframe}

\item{path}{path to output dir}
}
\value{
findings, dataframe with differences (if there are any)
}
\description{
Function to find differences in data
}
