% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{adm.fix_valuetype}
\alias{adm.fix_valuetype}
\title{Function to set the correct valuetype of data vs variables}
\usage{
adm.fix_valuetype(datafile, variables)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\value{
variables Return variables with corrected valuetypes
}
\description{
Function to set the correct valuetype of data vs variables
}
