% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{adm.table_get}
\alias{adm.table_get}
\title{Function to read a table from opal to R}
\usage{
adm.table_get(opal, projname, tablename, max_retries = 3, ...)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{projname}{Origin opal project name}

\item{tablename}{Origin opal table name}

\item{max_retries}{Integer: The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
datalist, list with datafiles & dictionary
}
\description{
Function to read a table from opal to R
}
