% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R, R/tables.R
\name{adm.table_save}
\alias{adm.table_save}
\title{Function to run a table save + table get + diffdf}
\usage{
adm.table_save(
  opal,
  projname,
  tablename,
  datafile,
  variables,
  categories = NULL,
  method = "write",
  diffdf = FALSE,
  max_retries = 3,
  ...
)

adm.table_save(
  opal,
  projname,
  tablename,
  datafile,
  variables,
  categories = NULL,
  method = "write",
  diffdf = FALSE,
  max_retries = 3,
  ...
)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{projname}{Origin opal project name}

\item{tablename}{Origin opal table name}

\item{datafile}{data dataframe}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}

\item{method}{String: write, update or overwrite}

\item{diffdf}{Boolean, Run a table_get() & diffdf after saving new table}

\item{max_retries}{Integer: The number of times R will try to read a datafile from, or write a datafile to, opal.}
}
\value{
variables Return variables with corrected valuetypes
}
\description{
Function to run a table save + table get + diffdf

Function to save a table from R to Opal
}
