% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{adm.table_save}
\alias{adm.table_save}
\title{Function to save a table from R to Opal}
\usage{
adm.table_save(
  opal,
  project,
  table,
  datafile,
  variables,
  categories = NULL,
  method,
  ...
)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{project}{Origin opal project name}

\item{table}{Origin opal table name}

\item{datafile}{data dataframe}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}

\item{method}{String: write, update or overwrite}
}
\description{
Function to save a table from R to Opal
}
