% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tres.R
\name{adm.tres_decryption}
\alias{adm.tres_decryption}
\title{Decrypt Data Using TRES}
\usage{
adm.tres_decryption(connection, values, ...)
}
\arguments{
\item{connection}{A connection object used to communicate with the TRES service.}

\item{values}{Column values to encrypt.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
Result as decrypted values
}
\description{
Decrypts data using the TRES decryption service.
}
