% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tres.R
\name{adm.tres_decryption}
\alias{adm.tres_decryption}
\title{Decrypt Data Columns Using TRES}
\usage{
adm.tres_decryption(connection, datafile, columns = NULL, ...)
}
\arguments{
\item{connection}{A connection object used to communicate with the TRES service.}

\item{datafile}{A data frame containing the encrypted data to be decrypted.}

\item{columns}{A character vector specifying which columns to decrypt. If NULL, columns starting with "3::" are selected.}
}
\value{
A data frame with decrypted values in the specified columns.
}
\description{
Decrypts specified columns in a data frame using the TRES decryption service.
}
