% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tres.R
\name{adm.tres_encryption}
\alias{adm.tres_encryption}
\title{Function to encrypt with tres}
\usage{
adm.tres_encryption(
  connection,
  datafile,
  columns = NULL,
  search_image = FALSE,
  ...
)
}
\arguments{
\item{connection}{Connection with RTRES}

\item{datafile}{data input}

\item{columns}{selected columns for encryption}
}
\value{
datafile, encrypted datafile
}
\description{
Function to encrypt with tres
}
