% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{adm.view_create}
\alias{adm.view_create}
\title{Function to create a view in Opal}
\usage{
adm.view_create(
  opal,
  view_projname,
  view_tablename,
  source,
  variables,
  categories = NULL,
  ...
)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{view_projname}{Origin opal project name}

\item{view_tablename}{Origin opal table name}

\item{source}{Vector with project and table names: project.table: CNSIM.CNSIM1}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}
}
\description{
Function to create a view in Opal
}
