% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{adm.view_update}
\alias{adm.view_update}
\title{Function to update a view in Opal}
\usage{
adm.view_update(
  opal,
  projname_view,
  tablename_view,
  source,
  variables,
  categories = NULL,
  ...
)
}
\arguments{
\item{opal}{a working opalr::opal_login}

\item{projname_view}{Origin opal project name}

\item{tablename_view}{Origin opal table name}

\item{source}{Vector with project and table names: project.table: CNSIM.CNSIM1}

\item{variables}{variables dataframe}

\item{categories}{categories dataframe}
}
\description{
Function to update a view in Opal
}
