% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.character_ids}
\alias{check.character_ids}
\title{Function to check if key is a character value}
\usage{
check.character_ids(datafile, id.name = "id", ...)
}
\arguments{
\item{datafile}{data input}

\item{id.name}{The name of the column representing the entity identifiers. Default is 'id'.}
}
\description{
Function to check if key is a character value
}
