% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.columns_cat}
\alias{check.columns_cat}
\title{Function to check if all columns listed in categories are also in the dataset}
\usage{
check.columns_cat(datafile, categories = NULL, ...)
}
\arguments{
\item{datafile}{data input}

\item{categories}{categories dataframe}
}
\description{
Function to check if all columns listed in categories are also in the dataset
}
