% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.columns_var}
\alias{check.columns_var}
\title{Function to check if all columns listed in variables are also in the dataset and vice-versa}
\usage{
check.columns_var(datafile, variables, ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check if all columns listed in variables are also in the dataset and vice-versa
}
