% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.datetime}
\alias{check.datetime}
\title{Function to check datetime}
\usage{
check.datetime(datafile, variables, format = "\%Y-\%m-\%d \%H:\%M:\%OS", ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}

\item{format}{format of dates (default = "\%Y-\%m-\%d \%H:\%M:\%OS")}
}
\description{
Function to check datetime
}
