% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.entitytype}
\alias{check.entitytype}
\title{Check Entity Type Consistency}
\usage{
check.entitytype(variables, ...)
}
\arguments{
\item{variables}{A data frame or list containing variable definitions.}
}
\value{
A warning if more than one entity type is present.
}
\description{
Ensures that only one entity type is defined across all variables.
}
