% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.infinite}
\alias{check.infinite}
\title{Check for Infinite Values in Data}
\usage{
check.infinite(datafile, ...)
}
\arguments{
\item{datafile}{A data frame containing the actual data.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A warning if any column contains infinite values.
}
\description{
Identifies columns containing infinite values (Inf or -Inf).
}
