% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.required_columns}
\alias{check.required_columns}
\title{Function to check presence of required columns labels, descriptions & encrypted in variables}
\usage{
check.required_columns(variables, ...)
}
\arguments{
\item{variables}{variables dataframe}
}
\description{
Function to check presence of required columns labels, descriptions & encrypted in variables
}
