% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.valuetype}
\alias{check.valuetype}
\title{Function to check valuetypes of data vs variables}
\usage{
check.valuetype(datafile, variables, ...)
}
\arguments{
\item{datafile}{data input}

\item{variables}{variables dataframe}
}
\description{
Function to check valuetypes of data vs variables
}
