% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.var_column_datatype}
\alias{check.var_column_datatype}
\title{Check Variable Object Datatypes}
\usage{
check.var_column_datatype(variables, ...)
}
\arguments{
\item{variables}{A data frame or list containing variable definitions.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A warning if any column is not of character type.
}
\description{
Ensures that all columns in the variables object are character type.
}
