% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.capture_logs}
\alias{.capture_logs}
\title{Function to capture all warnings & errors when running all check functions}
\usage{
.capture_logs(function_name)
}
\arguments{
\item{function_name}{Function to run & catch logs from}
}
\value{
logs Return list with warnings & errors
}
\description{
Function to capture all warnings & errors when running all check functions
}
