% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.capture_logs}
\alias{.capture_logs}
\title{Function to capture all warnings & errors when running all check functions}
\usage{
.capture_logs(run_function)
}
\arguments{
\item{run_function}{Function to run}
}
\value{
logs Return list with warnings & errors
}
\description{
Function to capture all warnings & errors when running all check functions
}
