% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.clean_NA}
\alias{.clean_NA}
\title{Function to cleanup NA or null or "" for diffdf}
\usage{
.clean_NA(datafile)
}
\arguments{
\item{datafile}{DataFrame to clean}
}
\value{
datafile, DataFrame with NA/null/"" cleaned
}
\description{
Function to cleanup NA or null or "" for diffdf
}
