% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.map_dtype}
\alias{.map_dtype}
\title{Function to capture all warnings & errors when running all check functions}
\usage{
.map_dtype(column)
}
\arguments{
\item{column}{Data column}
}
\description{
Function to capture all warnings & errors when running all check functions
}
