% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.map_dtype}
\alias{.map_dtype}
\title{Map R data types to Opal data types}
\usage{
.map_dtype(column)
}
\arguments{
\item{column}{A single R column object (e.g., vector)}
}
\value{
A string representing the Opal data type corresponding to the input column class
}
\description{
Translates R data types (e.g., integer, character) to corresponding Opal data types
(e.g., integer, text, decimal, boolean, datetime, date).
}
