% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.table_save_diffdf}
\alias{.table_save_diffdf}
\title{Function to save a table from R to Opal + a table_get() + diffdf}
\usage{
.table_save_diffdf(
  opal,
  projname,
  tablename,
  datafile,
  variables,
  categories = NULL,
  ...
)
}
\description{
Function to save a table from R to Opal + a table_get() + diffdf
}
